Switch BaseActor to an override
